Add debug flag to keep tempfiles after request

+++ lib/html_snapshot/screenshot_html_content.rb
@@ -41,11 +41,13 @@ module HTMLSnapshot
       gif_content = temporary_gif_file.read
 
       logger.debug(temporary_html_file_size: temporary_html_file.size, temporary_gif_file_size: temporary_gif_file.size)
-      temporary_html_file.close
-      temporary_html_file.unlink
+      unless ENV["KEEP_HTML_SNAPSHOT_TEMPFILES"]
+        temporary_html_file.close
+        temporary_html_file.unlink
 
-      temporary_gif_file.close
-      temporary_gif_file.unlink
+        temporary_gif_file.close
+        temporary_gif_file.unlink
+      end
 
       return gif_content
     end
